fix return type for day1

+++ day01/day01.hpp
@@ -3,7 +3,7 @@
 
 #include <string>
 
-int d01pt1(const std::filesystem::path& file_path);
-int d01pt2(const std::filesystem::path& file_path);
+int64_t d01pt1(const std::filesystem::path& file_path);
+int64_t d01pt2(const std::filesystem::path& file_path);
 
 #endif // DAY01_H
